Renamed TaskManager to TaskHandler for consistency
Moved TaskHandler to task package
