+++ src/main/java/yapper/Yapper.java
@@ -4,13 +4,14 @@ import java.util.Scanner;
 
 import yapper.instructions.InstructionHandler;
 import yapper.io.StringStorage;
+import yapper.tasks.TaskHandler;
 
 // Main function for Yapper
 public class Yapper {
     public static final int maxCapacity = 100;
 
     // Main ChatBot Loop
-    public static void startYappin(TaskManager taskManager) {
+    public static void startYappin(TaskHandler taskHandler) {
         System.out.println(StringStorage.START_UP_MESSAGE);
         System.out.println(StringStorage.HELP_MESSAGE); // TODO
 
@@ -20,7 +21,7 @@ public class Yapper {
             if (userInputString.trim().equals("bye")) {
                 break;
             }
-            InstructionHandler.handleInstruction(taskManager, userInputString);
+            InstructionHandler.handleInstruction(taskHandler, userInputString);
         }
 
         System.out.println(StringStorage.SHUT_DOWN_MESSAGE);
@@ -28,8 +29,8 @@ public class Yapper {
 
     public static void main(String[] args) {
         // Initialize
-        TaskManager taskManager = new TaskManager(maxCapacity);
+        TaskHandler taskHandler = new TaskHandler(maxCapacity);
         // Startup ChatBot Program
-        startYappin(taskManager);
+        startYappin(taskHandler);
     }
 }
