+++ src/main/java/yapper/instructions/InstructionHandler.java
@@ -1,6 +1,6 @@
 package yapper.instructions;
 
-import yapper.TaskManager;
+import yapper.tasks.TaskHandler;
 import yapper.Yapper;
 import yapper.exceptions.ErrorHandler;
 import yapper.exceptions.YapperException;
@@ -16,27 +16,27 @@ public class InstructionHandler {
     private static final int INDEX_OFFSET = 1;
 
     // UI Operations
-    public static void handleAddInstruction(TaskManager taskManager, Task task) {
+    public static void handleAddInstruction(TaskHandler taskHandler, Task task) {
         try {
-            ErrorHandler.checkIfListFull(taskManager.getCurrTaskTotal(), Yapper.maxCapacity);
+            ErrorHandler.checkIfListFull(taskHandler.getCurrTaskTotal(), Yapper.maxCapacity);
         } catch (YapperException e) {
             System.out.println(e.getMessage());
             return;
         }
-        taskManager.addTask(task);
+        taskHandler.addTask(task);
     }
-    public static void handleMarkingInstruction(TaskManager taskManager, Integer taskOrdinal, boolean isDone) {
+    public static void handleMarkingInstruction(TaskHandler taskHandler, Integer taskOrdinal, boolean isDone) {
         try {
-            ErrorHandler.checkIfTaskOrdinalWithinRange(taskManager.getCurrTaskTotal(), taskOrdinal);
+            ErrorHandler.checkIfTaskOrdinalWithinRange(taskHandler.getCurrTaskTotal(), taskOrdinal);
         } catch (YapperException e) {
             System.out.println(e.getMessage());
             return;
         }
-        taskManager.updateTaskStatus(taskOrdinal - INDEX_OFFSET, isDone);
+        taskHandler.updateTaskStatus(taskOrdinal - INDEX_OFFSET, isDone);
     }
 
 
-    public static void handleInstruction(TaskManager taskManager, String userInputString) {
+    public static void handleInstruction(TaskHandler taskHandler, String userInputString) {
         Instruction instruction = null;
         try {
             instruction = InputStringHandler.parseUserInput(userInputString);
@@ -48,32 +48,32 @@ public class InstructionHandler {
         switch (instructionType) {
         case TODO:
             String todoDesc = instruction.getInstructionDesc();
-            handleAddInstruction(taskManager,
+            handleAddInstruction(taskHandler,
                     new Todo(todoDesc) );
             break;
         case DEADLINE:
             String deadlineDesc = instruction.getInstructionDesc();
             String deadline = instruction.getTaskDates()[0];
-            handleAddInstruction(taskManager,
+            handleAddInstruction(taskHandler,
                     new Deadline(deadlineDesc, deadline) );
             break;
         case EVENT:
             String eventDesc = instruction.getInstructionDesc();
             String startDate = instruction.getTaskDates()[0];
             String endDate = instruction.getTaskDates()[1];
-            handleAddInstruction(taskManager,
+            handleAddInstruction(taskHandler,
                     new Event(eventDesc, startDate, endDate) );
             break;
         case LIST:
-            OutputStringHandler.printTasks(taskManager.getAllTasks(),
-                    taskManager.getCurrTaskTotal());
+            OutputStringHandler.printTasks(taskHandler.getAllTasks(),
+                    taskHandler.getCurrTaskTotal());
             break;
         case MARK:
-            handleMarkingInstruction(taskManager,
+            handleMarkingInstruction(taskHandler,
                     instruction.getTaskOrdinal(), true);
             break;
         case UNMARK:
-            handleMarkingInstruction(taskManager,
+            handleMarkingInstruction(taskHandler,
                     instruction.getTaskOrdinal(), false);
             break;
         } // BYE instruction is not handled here
